LAB 1 finish

+++ Web Programming/LAB1/LAB1/src/main/java/mk/finki/ukim/mk/lab/model/TicketOrder.java
@@ -9,5 +9,5 @@ public class TicketOrder {
     public String movieTitle;
     public String clientName;
     public String clientAddress;
-    public Long numberOfTickets;
+    public int numberOfTickets;
 }
